chore: reorganise storybook and add changeset
