+++ packages/remirror__react-ui/package.json
@@ -36,6 +36,7 @@
   "dependencies": {
     "@babel/runtime": "^7.22.3",
     "@mui/material": "^5.13.2",
+    "@mui/system": "^5.13.2",
     "@remirror/core": "3.0.0-beta.0",
     "@remirror/extension-blockquote": "3.0.0-beta.0",
     "@remirror/extension-bold": "3.0.0-beta.0",
