+++ packages/remirror__react-ui/src/providers/index.ts
@@ -1,1 +1,1 @@
-export * from './ui-theme-provider';
+export { useRemirrorDefaultMuiTheme } from './ui-theme-provider';
