+++ packages/remirror__react-ui/src/toolbar/base-toolbar.tsx
@@ -1,7 +1,7 @@
 import { Stack, StackProps } from '@mui/material';
 import React, { FC } from 'react';
 
-import { UiThemeProvider } from '../providers';
+import { UiThemeProvider } from '../providers/ui-theme-provider';
 
 export const Toolbar: FC<StackProps> = (props) => (
   <UiThemeProvider>
