+++ packages/storybook-react/.storybook/preview.js
@@ -3,7 +3,7 @@ import './storybook.css';
 export const parameters = {
   options: {
     storySort: {
-      order: ['Editors', 'Extensions', 'React Hooks', 'Components (labs)', 'Utils'],
+      order: ['Editors', 'Extensions', 'React Hooks', 'Components', 'React UI (labs)', 'Utils'],
     },
   },
 };
