+++ packages/storybook-react/src/tsconfig.json
@@ -62,10 +62,10 @@
       "path": "../../remirror__react/src"
     },
     {
-      "path": "../../remirror__react-ui/src"
+      "path": "../../remirror__react-editors/src"
     },
     {
-      "path": "../../remirror__react-editors/src"
+      "path": "../../remirror__react-ui/src"
     },
     {
       "path": "../../remirror__styles/src"
