+++ packages/storybook-react/stories/tsconfig.json
@@ -61,10 +61,10 @@
       "path": "../../remirror__react/src"
     },
     {
-      "path": "../../remirror__react-ui/src"
+      "path": "../../remirror__react-editors/src"
     },
     {
-      "path": "../../remirror__react-editors/src"
+      "path": "../../remirror__react-ui/src"
     },
     {
       "path": "../../remirror__styles/src"
